chore: Move B2C code to own folder
