+++ lib/b2c/magic_links.ts
