+++ lib/b2c/shared.ts
