+++ test/crypto_wallets.test.ts
@@ -1,4 +1,4 @@
-import { CryptoWallets } from "../lib/crypto_wallets";
+import { CryptoWallets } from "../lib/b2c/crypto_wallets";
 import { MOCK_FETCH_CONFIG, mockRequest } from "./helpers";
 
 jest.mock("../lib/shared");
