+++ test/helpers.ts
@@ -1,4 +1,4 @@
-import { request, requestConfig } from "../lib/shared";
+import { request, requestConfig } from "../lib/b2c/shared";
 import * as http from "http";
 
 export type Response = {
