+++ test/sessions.test.ts
@@ -1,6 +1,6 @@
 import * as jose from "jose";
-import { Sessions } from "../lib/sessions";
-import { ClientError } from "../lib/errors";
+import { Sessions } from "../lib/b2c/sessions";
+import { ClientError } from "../lib/b2c/errors";
 import { MOCK_FETCH_CONFIG, mockRequest } from "./helpers";
 
 function jwtConfig(projectID: string) {
