--- types/lib/envs.d.ts
@@ -1,2 +1,0 @@
-export declare const test = "https://test.stytch.com/v1/";
-export declare const live = "https://api.stytch.com/v1/";
