+++ types/lib/index.d.ts
@@ -1,4 +1,4 @@
-export { Client } from "./client";
-export * as envs from "./envs";
-export { UserSearchOperator } from "./users";
-export * from "./errors";
+export { Client } from "./b2c/client";
+export * as envs from "./b2c/envs";
+export { UserSearchOperator } from "./b2c/users";
+export * from "./b2c/errors";
